backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-spectacular==0.26.1
drf-writable-nested==0.7.0
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
install==1.3.5
jsonschema==4.17.3
Markdown==3.4.3
multidict==6.0.4
openai==0.27.4
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
sqlparse==0.4.3
tiktoken==0.3.3
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
yarg==0.1.9
yarl==1.8.2
gingerit=0.9.0